backend/requirements.txt
openai
fastapi
uvicorn
pillow
python-dotenv
python-multipart